// main-backend/client-dto/GetBulb.dto.ts
import { YeelightMode } from "./YeelightMode.enum";
import { YeelightModel } from "./YeelightModel.enum";
import { AvailableCommands } from "./AvailableCommands";

export interface GetBulbDtoPayload {
  bulbId: string;
}

export interface GetBulbDtoResponse {
  id: string;
  name?: string;
  model: YeelightModel;
  colorMode: YeelightMode;
  available_actions: AvailableCommands[];
  rgb: string;
  hue: number;
  sat: number;
  ct: number;
  power: boolean;
  status: boolean;
  bright: number;
}